Move worker-scanning into provisioning loop
