--- infrastructure/k8s/templates/taskcluster-worker-manager-deployment-workerscanner.yaml
@@ -1,48 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: taskcluster-worker-manager-workerscanner
-  labels: &ref_0
-    app.kubernetes.io/name: taskcluster-worker-manager
-    app.kubernetes.io/instance: '{{ .Release.Name }}'
-    app.kubernetes.io/component: taskcluster-worker-manager-workerscanner
-    app.kubernetes.io/part-of: taskcluster
-spec:
-  replicas: {{ int (.Values.worker_manager.procs.workerscanner.replicas) }}
-  selector:
-    matchLabels: *ref_0
-  template:
-    metadata:
-      annotations:
-        checksum/secret: '{{ include (print $.Template.BasePath "/taskcluster-worker-manager-secret.yaml") . | sha256sum }}'
-      labels: *ref_0
-    spec:
-      serviceAccountName: taskcluster-worker-manager
-      containers:
-        - name: taskcluster-worker-manager
-          image: '{{ .Values.dockerImage }}'
-          imagePullPolicy: Always
-          args:
-            - worker-manager/workerscanner
-          resources:
-            requests:
-              cpu: '{{ .Values.worker_manager.procs.workerscanner.cpu }}'
-              memory: '{{ .Values.worker_manager.procs.workerscanner.memory }}'
-          env:
-            - name: TASKCLUSTER_ROOT_URL
-              value: '{{ .Values.rootUrl }}'
-            - name: NODE_ENV
-              value: production
-          envFrom:
-            - secretRef:
-                name: taskcluster-worker-manager
-          ports: []
-          livenessProbe:
-            exec:
-              command:
-                - /bin/sh
-                - '-c'
-                - 'exit $(awk ''BEGIN{srand(); print (rand() > 0.3)}'')'
-            initialDelaySeconds: 86400
-            periodSeconds: 60
-            failureThreshold: 1
