+++ infrastructure/k8s/values.yaml
@@ -191,10 +191,6 @@ worker_manager:
       replicas: 1
       cpu: 50m
       memory: 100Mi
-    workerscanner:
-      replicas: 1
-      cpu: 50m
-      memory: 100Mi
     expire_workers:
       cpu: 50m
       memory: 100Mi
