+++ services/worker-manager/Procfile
@@ -2,7 +2,6 @@
 
 web: cd services/worker-manager && node src/main.js server
 provisioner: cd services/worker-manager && node src/main.js runProvisioner
-workerscanner: cd services/worker-manager && node src/main.js workerScanner
 expire-workers: cd services/worker-manager && node src/main.js expireWorkers
 expire-worker-pools: cd services/worker-manager && node src/main.js expireWorkerPools
 expire-errors: cd services/worker-manager && node src/main.js expireErrors
