+++ ui/docs/manual/deploying/monitoring.mdx
@@ -45,7 +45,6 @@ have their schedules and maximum allowed duration defined here. All times are re
 | web-server     | cleanup-expire-auth-codes    | taskcluster.web-server     | 86400              | At 12:00 AM |
 | web-server     | cleanup-expire-access-tokens | taskcluster.web-server     | 86400              | At 12:00 AM |
 | worker-manager | provisioner                  | taskcluster.worker-manager | continuous         | continuous  |
-| worker-manager | workerscanner                | taskcluster.worker-manager | continuous         | continuous  |
 | worker-manager | expire-workers               | taskcluster.worker-manager | 86400              | At 12:00 AM |
 | worker-manager | expire-worker-pools          | taskcluster.worker-manager | 86400              | At 01:00 AM |
 | worker-manager | expire-errors                | taskcluster.worker-manager | 86400              | At 12:10 AM |
